Show ConfirmationDialog when receiving DLC.broadcast state change
